Completare la funzione checkCollide

+++ src/dataObstacles/Sbarra.java
@@ -49,10 +49,10 @@ public class Sbarra extends Ostacolo{
 	public void draw( Graphics g ) throws SlickException
 		{ immagine.draw( ostr.getX(), ostr.getY(), ostr.getWidth(), ostr.getHeight() ); }
 	
-	public float getX()
+	public int getX()
 		{ return (int)ostr.getX(); }
 
-	public float getY()
+	public int getY()
 		{ return (int)ostr.getY(); }
 	
 	public float getWidth()
@@ -64,7 +64,7 @@ public class Sbarra extends Ostacolo{
 	public boolean contains( int x, int y )
 		{ return ostr.contains( x, y ); }
 	
-	public void setXY( float x, float y, String function )
+	public void setXY( int x, int y, String function )
 		{			
 			if(function.compareTo( "move" ) == 0)
 				ostr.setLocation( ostr.getX() + x, ostr.getY() + y );
